requirements.txt
beautifulsoup4==4.10.0
numpy==1.21.0
pandas==1.3.0
python-decouple==3.6
requests==2.23.0
SQLAlchemy==1.4.31
SQLAlchemy_Utils==0.38.2
psycopg2==2.9.3
